### 2024-08-06 d270540
feat: compliance with multiple river estimator checks
