### 2024-08-06 6fdb721
fix: scipy version (latest removed neccesary func)
