+++ setup.py
@@ -18,7 +18,7 @@ base_packages = [
     "pandas~=2.2.2",
     "numpy~=1.26.4",
     "river~=0.21.1",
-    "scipy~=1.13.0"
+    "scipy~=1.13.0",
     "tqdm~=4.66.4",
     "ordered-set~=4.1.0",
     "torchviz~=0.0.2",
