+++ deep_river/regression/multioutput.py
@@ -2,8 +2,8 @@ import typing
 from collections import OrderedDict
 from typing import Callable, Type, Union
 
-from ordered_set import OrderedSet
 import torch
+from ordered_set import OrderedSet
 from river.base import MultiTargetRegressor as RiverMultiTargetRegressor
 from river.base.typing import FeatureName, RegTarget
 
