+++ deep_river/classification/zoo.py
@@ -1,7 +1,7 @@
 from typing import Callable, Union
 
-from torch import nn
 import torch
+from torch import nn
 
 from deep_river.classification import Classifier
 
