+++ deep_river/classification/classifier.py
@@ -1,5 +1,6 @@
-from typing import Callable, Dict, List, Type, Union
 import pdb
+from typing import Callable, Dict, List, Type, Union
+
 import pandas as pd
 import torch
 from ordered_set import OrderedSet
