+++ deep_river/utils/tensor_conversion.py
@@ -187,7 +187,8 @@ def output2proba(
     # Add opposite class if first label seems to be binary
     elif classes == [True] or classes == [False]:
         all_classes = classes | ([not classes[0]])
-    # Add `unobserved` classes if pred shape is greater than number of observed classes
+    # Add `unobserved` classes if pred shape is greater than number of
+    # observed classes
     else:
         all_classes = classes | (
             [f"Unobserved{i}" for i in range(preds_np.shape[1] - len(classes))]
