+++ deep_river/classification/classifier.py
@@ -1,5 +1,4 @@
-import pdb
-from typing import Callable, Dict, List, Type, Union
+from typing import Callable, Dict, Type, Union
 
 import pandas as pd
 import torch
