### 2024-08-12 #7076 @kulbachcedric
@cheginit thanks for editing!

I would suggest the following reviewers:

- eminizer (seems to be a good fit according to her libraries)
- mbdemoraes (according to his interests in machine learning in data streams) and maybe
- Karangupta1994 (also according to the interests stated within the [joss reviewers](https://reviewers.joss.theoj.org/reviewers/search?language=26&area_id=5&keywords=stream&name=&commit=Update+search)
- JHoelli is also familiar with this kind of topic, however she worked with us in one group and I think it would be biased.

Again many thanks in advance and kind regards
Cedric
