### 2024-09-11 #7076 @kulbachcedric
Hi @cheginit,
i had a short chat with @musabgultekin.
He would be interested in reviewing our JOSS paper.

Best
Cedric
