### 2024-09-11 #7076 @musabgultekin
Hi @cheginit

Yes, I can review the paper for the Journal of Open Source Software.
I've built open-source machine learning project and I would love to help on this submission.
And as you informed, I can start within the next two weeks. I'll be active and I'll openly conduct the process.
