### 2024-09-14 #7226 @editorialbot
[REVIEW]: DeepRiver: A Deep Learning Library for Data Streams
**Submitting author:** @kulbachcedric (<a href="http://orcid.org/0000-0002-9363-4728">Cedric Kulbach</a>)
**Repository:** https://github.com/online-ml/deep-river
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.8
