+++ deep_river/preprocessing/__init__.py
@@ -1,0 +1,6 @@
+from deep_river.preprocessing.embedding import EmbeddingTransformer
+from .__version__ import __version__  # noqa: F401
+
+__all__ = [
+    "EmbeddingTransformer"
+]
