### 2024-10-10 #7226 @musabgultekin
@kulbachcedric

Okay I find that python 3.12 and 3.13 is not supported. Could you clarify that in the installation or add support for 3.12 at least?
