### 2024-10-10 #7226 @musabgultekin
Also it would have been better if you could publish a new release with the latest commit. So we don't need to install with git. As the current "pip install deep-river" doesn't include all commits
