### 2024-10-10 #7226 @musabgultekin
@kulbachcedric

Regarding the paper:

- It lacks a section dedicated to comparing DeepRiver with other existing tools or libraries that address deep learning in streaming or online learning contexts. So we need to update the paper to include related work. Such as scikit-multiflow, Apache SAMOA etc
- Also I see that citations are sometimes inconsistently formatted. could you format the citations consistently? See: https://pandoc.org/MANUAL.html#extension-citations
