### 2024-11-05 #7226 @kulbachcedric
Hi @all,

sorry for the late response, I was looking into the wrong thread....
However, SAMOA was not related to deep-river when it was created and focusses more on the integration of Python into the MOA Framework. The Package Sci-kit multiflow merged in 2021 with the package creme and was rebranded into river. In this case Sci-kit multiflow is not related any more and does not cover the deep learning part.

Regarding the consistancy, I will update the citations within the paper branch of the deep-river repository.
Is that ok for you?

Thanks again for reviewing and kind regards
Cedric
