### 2024-11-06 online-ml/deep-river#95 @atanikan
Updates to installation
Hello! As a reviewer for JOSS, I’d like to suggest a few updates and additions to the software installation process. For simplicity, I’ve consolidated them here in a single issue.

* Python Environment Setup: It would be helpful to specify the minimum required Python version and include steps for setting up either a virtual environment or a conda environment for package installation. This will make the installation process smoother for users.

* Link to Examples: I recommend adding a link to the [examples folder](https://github.com/online-ml/deep-river/blob/master/docs/examples) in your README or documentation. This can help users quickly locate example scripts and adapt or modify them as needed.

* Jupyter Dependency: Since the examples utilize Jupyter notebooks, it would be beneficial to list jupyter as a dependency when installing the package via pip. I noticed it’s missing from the pip list ((deep-river-joss) ADITYAs-MacBook-Pro-2:classification adityatanikanti$ pip list | grep "jupyter").

* Broken Example: The example for the Simple Classification Model in this [notebook](https://github.com/online-ml/deep-river/blob/master/docs/examples/classification/example_classification.ipynb) isn’t working, as well as some other examples. The error output is as follows:

```bash
Copy code
0it [00:00, ?it/s]
Cell In[19], line 6
      4 metric = metrics.Accuracy()
      5 for x, y in tqdm(dataset.take(5000)):
----> 6     y_pred = model_pipeline.predict_one(x)  # make a prediction
      7     metric = metric.update(y, y_pred)  # update the metric
      8     model_pipeline = model_pipeline.learn_one(x, y)  # make the model learn

AttributeError: 'NoneType' object has no attribute 'predict_one'
```
Please let me know if any further clarification is needed on these points. Thank you!
