### 2024-11-06 #7226 @atanikan
Hi @kulbachcedric

I have opened an issue in your repos issues tracker [#95](https://github.com/online-ml/deep-river/issues/95). Please let me know if you need clarification.
