+++ deep_river/preprocessing/embedding.py
@@ -1,84 +1,145 @@
-from typing import Dict
-
-import numpy as np
-import river.base as base
 import torch
-from transformers import AutoModel, AutoTokenizer
+import torch.nn as nn
+import torch.optim as optim
+from river import base
+from torchtext.vocab import build_vocab_from_iterator
+from collections import defaultdict
 
 
 class EmbeddingTransformer(base.Transformer):
-    """
-    A class for processing text input and generating embeddings using a pre-trained transformer model.
-
-    This class leverages the Hugging Face Transformers library to load a specified model and its tokenizer.
-    It can concatenate text from input dictionaries, tokenize the text, and generate embeddings, which can
-    be utilized for various natural language processing tasks such as semantic similarity, clustering, etc.
-
-    Attributes:
-        model_name (str): The name of the pre-trained transformer model to be used.
-    """
-
-    def __init__(self, model_name: str = "bert-base-uncased"):
+    def __init__(self, vocab_size, embedding_dim, tokenizer, learning_rate=0.01):
         """
-        Initializes the EmbeddingTransformer with the specified model and tokenization length.
+        Transformer that generates word embeddings and transforms variable-length input into a
+        fixed-size vector using mean pooling. The input is a dictionary of sentences/words,
+        which are tokenized and transformed into fixed-size vectors.
 
-        Args:
-            model_name (str): The name of the pre-trained transformer model to use (default is "bert-base-uncased").
+        Parameters
+        ----------
+        vocab_size : int
+            Size of the vocabulary.
+        embedding_dim : int
+            Dimensionality of the word embeddings.
+        tokenizer : callable
+            Function that tokenizes a sentence/word into tokens.
+        learning_rate : float
+            Learning rate for the optimizer.
         """
-        self.tokenizer = AutoTokenizer.from_pretrained(model_name)
-        self.model = AutoModel.from_pretrained(model_name)
-        self.model.eval()
+        self.vocab_size = vocab_size
+        self.embedding_dim = embedding_dim
+        self.tokenizer = tokenizer  # Tokenizer that converts sentences/words into tokens
+        self.learning_rate = learning_rate
 
-    def _concat_dict_values(self, x: Dict) -> str:
+        # PyTorch model
+        self.model = WordEmbeddingModel(vocab_size, embedding_dim)
+        self.optimizer = optim.Adam(self.model.parameters(), lr=learning_rate)
+        self.criterion = nn.CrossEntropyLoss()
+
+        # Call the PyTorch model's train mode
+        self.model.train()
+
+    def learn_one(self, x):
         """
-        Helper function to concatenate all values from the input dictionary into a single string.
+        Update the model with a single instance (x).
 
-        Args:
-            x (dict): A dictionary containing values to be concatenated.
+        Parameters
+        ----------
+        x : dict
+            Dictionary where the keys are words/sentences and values are text (strings).
 
-        Returns:
-            str: A single string composed of all values in the dictionary, separated by spaces.
+        Returns
+        -------
+        self
         """
-        return " ".join(str(value) for value in x.values())
+        # Tokenize and concatenate token lists from the dictionary
+        token_list = []
+        for text in x.values():
+            token_list.extend(self.tokenizer(text))  # Tokenize each sentence/word
 
-    def transform_one(self, x: Dict) -> Dict:
+        input_tensor = torch.tensor([token_list], dtype=torch.long)
+
+        # Forward pass
+        embedding = self.model(input_tensor)
+
+        # Loss computation and backward pass
+        loss = self.criterion(embedding, input_tensor)  # Target and input should be aligned for the loss
+        self.optimizer.zero_grad()
+        loss.backward()
+        self.optimizer.step()
+
[…]
