+++ deep_river/preprocessing/__init__.py
@@ -1,6 +1,4 @@
-from deep_river.preprocessing.embedding import EmbeddingTransformer
-from .__version__ import __version__  # noqa: F401
+from deep_river.preprocessing.embedding import (EmbeddingTransformer,
+                                                WordEmbeddingModel)
 
-__all__ = [
-    "EmbeddingTransformer"
-]
+__all__ = ["EmbeddingTransformer", "WordEmbeddingModel"]
