### 2024-11-10 c9a3bdd
Add link to the examples folder
