### 2024-11-10 #7226 @kulbachcedric
> Hi @kulbachcedric
>
> I have opened an issue in your repos issues tracker [#95](https://github.com/online-ml/deep-river/issues/95). Please let me know if you need clarification.

Hi @atanikan,
I'll work on the mentioned points within the branch [95-updates-to-installation](https://github.com/online-ml/deep-river/tree/95-updates-to-installation). In the upcoming week there will be a PR :-)

> @kulbachcedric Could you mind adding a separate section for the related work that you think of appropriate related work. Also I'm waiting update on the other comments I made above. Thank you!

Hi @musabgultekin,
good point, I'll also work on that in the upcoming week within the [paper](https://github.com/online-ml/deep-river/tree/paper) branch. Based on your points already mentioned I made a few smaller changes within this branch.
Is it in the way you thought of?

Thanks both of you for your feedback!
