+++ deep_river/anomaly/ae.py
@@ -68,7 +68,7 @@ class Autoencoder(DeepEstimator, AnomalyDetector):
     >>> from river.preprocessing import MinMaxScaler
 
     >>> dataset = CreditCard().take(5000)
-    >>> metric = metrics.ROCAUC(n_thresholds=50)
+    >>> metric = metrics.RollingROCAUC(window_size=5000)
 
     >>> class MyAutoEncoder(torch.nn.Module):
     ...     def __init__(self, n_features, latent_dim=3):
