+++ deep_river/anomaly/probability_weighted_ae.py
@@ -62,7 +62,7 @@ class ProbabilityWeightedAutoencoder(ae.Autoencoder):
 
     >>> _ = manual_seed(42)
     >>> dataset = CreditCard().take(5000)
-    >>> metric = metrics.ROCAUC(n_thresholds=50)
+    >>> metric = metrics.RollingROCAUC(window_size=5000)
 
     >>> class MyAutoEncoder(torch.nn.Module):
     ...     def __init__(self, n_features, latent_dim=3):
