+++ README.md
@@ -135,7 +135,7 @@ MicroAverage(MAE): 34.31
 >>> from river.preprocessing import MinMaxScaler
 
 >>> dataset = CreditCard().take(5000)
->>> metric = metrics.ROCAUC(n_thresholds=50)
+>>> metric = metrics.RollingROCAUC(window_size=5000)
 
 >>> class MyAutoEncoder(nn.Module):
 ...     def __init__(self, n_features, latent_dim=3):
@@ -160,8 +160,8 @@ MicroAverage(MAE): 34.31
 ...     model.learn_one(x=x)
 ...     metric.update(y, score)
 ...
->>> print(f"ROCAUC: {metric.get():.4f}")
-ROCAUC: 0.7812
+>>> print(f"Rolling ROCAUC: {metric.get():.4f}")
+Rolling ROCAUC: 0.8901
 
 ```
 
