### 2024-11-18 f5dac54
Refactor paper
