+++ paper/paper.md
@@ -37,8 +37,8 @@ However, as data evolves over time, it becomes crucial to adapt machine learning
 This adaptation is achieved through online learning or continuous ML technologies.
 Although deep learning technologies have demonstrated outstanding performance on predefined datasets, their application to online, streaming, and continuous learning scenarios has been limited.
 
-[`DeepRiver`](https://github.com/online-ml/deep-river) is a Python package for deep learning on data streams.
-Built on top of [`River`](https://riverml.xyz/latest/)[@montiel2021river] and [`PyTorch`](https://pytorch.org)[@paszke2017automatic], it offers a unified API for both supervised and unsupervised learning.
+`DeepRiver`[@githubGitHubOnlinemldeepriver] is a Python package for deep learning on data streams.
+Built on top of River[@montiel2021river] and PyTorch[@paszke2017automatic], it offers a unified API for both supervised and unsupervised learning.
 Additionally, it provides a suite of tools for preprocessing data streams and evaluating deep learning models.
 
 # Statement of need
@@ -48,32 +48,71 @@ The acceleration facilitated by these algorithms necessitates constant adaptatio
 To address this imperative, adopting online learning and continuous ML technologies becomes paramount.
 While deep learning technologies have demonstrated exceptional performance on static, predefined datasets, their application to dynamic and continuously evolving data streams remains underexplored.
 The absence of widespread integration of deep learning into online, streaming, and continuous learning scenarios hampers the full potential of these advanced algorithms in real-time decision-making [@kulbach2024retrospectivetutorialopportunitieschallenges].
-The emergence of the [`DeepRiver`](https://github.com/online-ml/deep-river) Python package fills a critical void in the field of deep learning on data streams.
-Leveraging the capabilities of [`River`](https://riverml.xyz/latest/)[@montiel2021river] and [`PyTorch`](https://pytorch.org)[@paszke2017automatic], `DeepRiver` offers a unified API for both supervised and unsupervised learning, providing a seamless bridge between cutting-edge deep learning techniques and the challenges posed by dynamic data streams.
+The emergence of the `DeepRiver`[@githubGitHubOnlinemldeepriver] Python package fills a critical void in the field of deep learning on data streams.
+Leveraging the capabilities of River[@montiel2021river] and PyTorch[@paszke2017automatic], `DeepRiver` offers a unified API for both supervised and unsupervised learning, providing a seamless bridge between cutting-edge deep learning techniques and the challenges posed by dynamic data streams.
 Moreover, the package equips practitioners with essential tools for data stream preprocessing and the evaluation of deep learning models in dynamic, real-time environments.
 This was already made use of in the context of Streaming Anomaly Detection [@cazzonelli2022detecting].
 As the demand for effective and efficient adaptation of machine learning systems to evolving data structures continues to grow, the integration of `DeepRiver` into the landscape becomes crucial.
 This package stands as a valuable asset, unlocking the potential for deep learning technologies to excel in online, streaming, and continuous learning scenarios.
 The need for such advancements is evident in the quest to harness the full power of machine learning in dynamically changing environments, ensuring our decision-making processes remain accurate, relevant, and agile in the face of evolving data landscapes.
 
+# Related Work
+
+Online machine learning involves updating models incrementally as new data arrives, rather than retraining models from scratch. 
+Several frameworks and libraries have been developed to support this paradigm:
+
+* scikit-multiflow [@10.5555/3291125.3309634]
+    * Python-based Library: Inspired by the Java-based MOA framework, designed for streaming data and online learning in Python. 
+    * Key Features:
+      * Supports algorithms like Hoeffding Trees, online bagging, and boosting. 
[…]
