+++ paper/paper.bib
@@ -56,3 +56,55 @@
   publisher={IEEE},
   doi={10.1109/TKDE.2018.2876857}
 }
+@article{10.5555/1756006.1859903,
+author = {Bifet, Albert and Holmes, Geoff and Kirkby, Richard and Pfahringer, Bernhard},
+title = {MOA: Massive Online Analysis},
+year = {2010},
+issue_date = {3/1/2010},
+publisher = {JMLR.org},
+volume = {11},
+issn = {1532-4435},
+abstract = {Massive Online Analysis (MOA) is a software environment for implementing algorithms and running experiments for online learning from evolving data streams. MOA includes a collection of offline and online methods as well as tools for evaluation. In particular, it implements boosting, bagging, and Hoeffding Trees, all with and without Na\"{\i}ve Bayes classifiers at the leaves. MOA supports bi-directional interaction with WEKA, the Waikato Environment for Knowledge Analysis, and is released under the GNU GPL license.},
+journal = {J. Mach. Learn. Res.},
+month = aug,
+pages = {1601–1604},
+numpages = {4}
+}
+@article{10.5555/3291125.3309634,
+author = {Montiel, Jacob and Read, Jesse and Bifet, Albert and Abdessalem, Talel},
+title = {Scikit-multiflow: a multi-output streaming framework},
+year = {2018},
+issue_date = {January 2018},
+publisher = {JMLR.org},
+volume = {19},
+number = {1},
+issn = {1532-4435},
+abstract = {scikit-multiflow is a framework for learning from data streams and multi-output learning in Python. Conceived to serve as a platform to encourage the democratization of stream learning research, it provides multiple state-of-the-art learning methods, data generators and evaluators for different stream learning problems, including single-output, multi-output and multi-label. scikit-multiflow builds upon popular open source frameworks including scikit-learn, MOA and MEKA. Development follows the FOSS principles. Quality is enforced by complying with PEP8 guidelines, using continuous integration and functional testing. The source code is available at https://github.com/scikit-multiflow/scikit-multiflow.},
+journal = {J. Mach. Learn. Res.},
+month = jan,
+pages = {2915–2914},
+numpages = {5},
+keywords = {drift detection, machine learning, multi-output, python, stream data}
+}
+@software{creme,
+  title = {{creme}, a {P}ython library for online machine learning},
+  author = {Halford, Max and Bolmier, Geoffrey and Sourty, Raphael and Vaysse, Robin and Zouitine, Adil},
+  url = {https://github.com/MaxHalford/creme},
+  version = {0.6.1},
+  date = {2020-06-10},
+  year = {2019}
+}
+@misc{capymoaCapyMOAx2024,
+	author = {},
+	title = {{C}apy{M}{O}{A} \&#x2014; {C}apy{M}{O}{A} --- capymoa.org},
+	howpublished = {\url{https://capymoa.org}},
+	year = {},
+	note = {[Accessed 18-11-2024]},
+}
+@misc{githubGitHubOnlinemldeepriver,
+	author = {},
+	title = {{G}it{H}ub - online-ml/deep-river --- github.com},
+	howpublished = {\url{https://github.com/online-ml/deep-river}},
+	year = {},
+	note = {[Accessed 18-11-2024]},
+}
