### 2024-11-18 #7226 @kulbachcedric
Hi @all
I refactored the [paper branch](https://github.com/online-ml/deep-river/tree/paper).
If there are any changes required, just let me know.

Thanks again for reviewing!

Best
Cedric
