### 2024-12-01 f0fb6a9
remove furhter authors as there is no mail
