### 2024-12-04 #7226 @musabgultekin
Hi @kulbachcedric .

- Related work section should be added to the paper. I don't see this in the paper.
- CONTRIBUTING file.
- Release to the pip with the latest commits

Only these are my concerns at the moment. Other than that it looks good.
