### 2024-12-13 #7226 @hoanganhngo610
Hi @musabgultekin! I am Hoang-Anh, one of the co-authors of the paper.

Regarding your first comment that "Related work section should be added to the paper", I believe that @kulbachcedric has already assessed this within his latest changes to the paper. The part [Related Work](https://github.com/online-ml/deep-river/blob/paper/paper/paper.md#related-work) has mentioned other frameworks within the field of ML, including `scikit-multiflow` and `creme`, the two predecessors of `River`, and `MOA` in Java and `capyMOA` as a Python interface for MOA. However, we also noted that `DeepRiver` differentiates itself by intergrating deep learning capabilities into streaming data workflow, allowing neural network models to continuously learn.

I hope that the content of this part would be satisfactory to you. Regarding the addition of CONTRIBUTING file and release the latest commits to pip, we will finish that ASAP.

Best regards,
Hoang-Anh
