+++ setup.cfg
@@ -1,4 +1,4 @@
 [flake8]
 exclude = docs
-max-line-length = 79
+max-line-length = 100
 ignore = E203 W503
