### 2024-12-24 90b1467
add pandas stubs
