+++ deep_river/anomaly/__init__.py
@@ -1,11 +1,8 @@
 from .ae import Autoencoder
 from .probability_weighted_ae import ProbabilityWeightedAutoencoder
 from .rolling_ae import RollingAutoencoder
-from .scaler import (
-    AnomalyMeanScaler,
-    AnomalyMinMaxScaler,
-    AnomalyStandardScaler,
-)
+from .scaler import (AnomalyMeanScaler, AnomalyMinMaxScaler,
+                     AnomalyStandardScaler)
 
 """
 This module contains the anomaly detection algorithms for the
