+++ deep_river/anomaly/probability_weighted_ae.py
@@ -114,9 +114,7 @@ class ProbabilityWeightedAutoencoder(ae.Autoencoder):
         )
         self.window_size = window_size
         self.skip_threshold = skip_threshold
-        self.rolling_mean = utils.Rolling(
-            stats.Mean(), window_size=window_size
-        )
+        self.rolling_mean = utils.Rolling(stats.Mean(), window_size=window_size)
         self.rolling_var = utils.Rolling(stats.Var(), window_size=window_size)
 
     def learn_one(
@@ -165,10 +163,7 @@ class ProbabilityWeightedAutoencoder(ae.Autoencoder):
 
         loss_scaled = (losses_numpy - mean) / math.sqrt(var)
         prob = ndtr(loss_scaled)
-        loss = (
-            torch.tensor((self.skip_threshold - prob) / self.skip_threshold)
-            * loss
-        )
+        loss = torch.tensor((self.skip_threshold - prob) / self.skip_threshold) * loss
 
         self.optimizer.zero_grad()
         loss.backward()
