+++ benchmarks/render.py
@@ -3,9 +3,9 @@ import shutil
 from pathlib import Path
 from typing import List
 
+import pandas as pd
 from dominate.tags import pre
 from watermark import watermark
-import pandas as pd
 
 
 def render_df(df_path: Path) -> dict:
