+++ deep_river/classification/__init__.py
@@ -1,7 +1,6 @@
 from deep_river.classification.classifier import Classifier
 from deep_river.classification.rolling_classifier import RollingClassifier
-from deep_river.classification.zoo import (LogisticRegression,
-                                           MultiLayerPerceptron)
+from deep_river.classification.zoo import LogisticRegression, MultiLayerPerceptron
 
 """
 This module contains the classifiers for the deep_river package.
