+++ deep_river/utils/__init__.py
@@ -2,8 +2,14 @@
 
 from .estimator_checks import check_estimator
 from .params import get_activation_fn, get_init_fn, get_loss_fn, get_optim_fn
-from .tensor_conversion import (deque2rolling_tensor, df2tensor, dict2tensor,
-                                float2tensor, labels2onehot, output2proba)
+from .tensor_conversion import (
+    deque2rolling_tensor,
+    df2tensor,
+    dict2tensor,
+    float2tensor,
+    labels2onehot,
+    output2proba,
+)
 
 __all__ = [
     "check_estimator",
