+++ pyproject.toml
@@ -127,9 +127,9 @@ all = [
 # Add any console scripts if applicable
 # deep-river-cli = "deep_river.cli:main"
 
-[[tool.mypy.overrides]]
-module = ["untyped_package.*"]
-follow_untyped_imports = true
+[tool.mypy]
+ignore_missing_imports = true
 
 [tool.isort]
 profile = "black"
+
