### 2024-12-24 96c5d27
Add poetry and CONTRIBUTING.md
