+++ CONTRIBUTING.md
@@ -1,0 +1,182 @@
+# Contribution guidelines
+
+## What to work on?
+
+We have a [public roadmap](https://github.com/orgs/online-ml/projects/3) that lists what has been done, what we're currently doing, and what needs doing. There's also an icebox with high level ideas that need framing. You're welcome to pick anything that takes your fancy and that you deem important. Feel free to [open a discussion](https://github.com/online-ml/river/discussions/new) if you want to clarify a topic and/or want to be formally assigned a task in the board.
+
+Of course, you're welcome to propose and contribute new ideas. We encourage you to [open a discussion](https://github.com/online-ml/river/discussions/new) so that we can align on the work to be done. It's generally a good idea to have a quick discussion before opening a pull request that is potentially out-of-scope.
+
+## Fork/clone/pull
+
+The typical workflow for contributing to River is:
+
+1. Fork the `main` branch from the [GitHub repository](https://github.com/online-ml/deep-river/).
+2. Clone your fork locally.
+3. Commit changes.
+4. Push the changes to your fork.
+5. Send a pull request from your fork back to the original `main` branch.
+
+## Local setup
+
+Start by cloning the repository:
+
+```sh
+git clone https://github.com/online-ml/deep-river
+```
+
+Next, you'll need a Python environment. A nice way to manage your Python versions is to use pyenv, which can installed [here](https://github.com/pyenv/pyenv-installer). Once you have pyenv, you can install the latest Python version River supports:
+
+```sh
+pyenv install -v $(cat .python-version)
+```
+
+```sh
+curl -sSL https://install.python-poetry.org | python3 -
+```
+
+Now you're set to install River and activate the virtual environment:
+
+```sh
+poetry install
+poetry shell
+```
+
+Finally, install the [pre-commit](https://pre-commit.com/) push hooks. This will run some code quality checks every time you push to GitHub.
+
+```sh
+pre-commit install --hook-type pre-push
+```
+
+You can optionally run `pre-commit` at any time as so:
+
+```sh
+pre-commit run --all-files
+```
+
+## Making changes
+
+You're now ready to make some changes. We strongly recommend that you to check out River's source code for inspiration before getting into the thick of it. How you make the changes is up to you of course. However we can give you some pointers as to how to test your changes. Here is an example workflow that works for most cases:
+
+- Create and open a Jupyter notebook at the root of the directory.
+- Add the following in the code cell:
+
+```py
+%load_ext autoreload
+%autoreload 2
+```
+
+- The previous code will automatically reimport River for you whenever you make changes.
+- For instance, if a change is made to `linear_model.LinearRegression`, then rerunning the following code doesn't require rebooting the notebook:
+
+```py
+from river import linear_model
+
+model = linear_model.LinearRegression()
+```
+
+## Creating a new estimator
+
+1. Pick a base class from the `base` module.
+2. Check if any of the mixin classes from the `base` module apply to your implementation.
+3. Make you've implemented the required methods, with the following exceptions:
+   1. Stateless transformers do not require a `learn_one` method.
+   2. In case of a classifier, the `predict_one` is implemented by default, but can be overridden.
+4. Add type hints to the parameters of the `__init__` method.
+5. If possible provide a default value for each parameter. If, for whatever reason, no good default exists, then implement the `_unit_test_params` method. This is a private method that is meant to be used for testing.
+6. Write a comprehensive docstring with example usage. Try to have empathy for new users when you do this.
[…]
