### 2024-12-24 9354d76
Add poetry and CONTRIBUTING.md
