+++ pyproject.toml
@@ -19,7 +19,6 @@ classifiers = [
     "License :: OSI Approved :: BSD License",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -28,7 +27,7 @@ classifiers = [
     "Programming Language :: Python :: Implementation :: PyPy",
 ]
 [tool.poetry.dependencies]
-python = ">=3.8"
+python = "<4.0,>=3.9"
 scikit-learn = "~1.5.0"
 torch = "~2.2.2"
 pandas = "~2.2.2"
@@ -52,7 +51,6 @@ black = ">=24.8.0"
 flake8 = ">=7.1.1"
 isort = ">=5.13.2"
 jupyter = ">=1.0.0"
-pandas-stubs = ">=2.2.3"
 pyupgrade = "==3.17.0"
 flask = ">=3.0.2"
 ipykernel = ">=6.9.0"
