+++ pyproject.toml
@@ -66,6 +66,7 @@ mkdocs-jupyter = ">=0.20.0"
 nbconvert = ">=6.4.2"
 jupyter_contrib_nbextensions = "0.7.0"
 numpydoc = ">=1.2"
+notebook = "==6.4.3"
 spacy = ">=3.2.2"
 jinja2 = ">=3.0.3"
 dominate = "*"
