+++ .github/workflows/unit-tests.yml
@@ -35,6 +35,8 @@ jobs:
 
       - name: Install Python dependencies
         run: |
+          python -m pip install --upgrade pip
+          pip install poetry
           poetry install --extra dev
 
       - name: Download datasets
