+++ .github/workflows/unit-tests.yml
@@ -37,6 +37,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install poetry
+          poetry install 
           poetry install --with dev
 
       - name: Download datasets
