+++ .github/workflows/mkdocs.yml
@@ -30,8 +30,9 @@ jobs:
       - name: Install Python dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install wheel
-          pip install -e ".[dev]"
+          pip install poetry
+          poetry install 
+          poetry install --with dev
       - name: Build docs
         run: make doc
 
