### 2024-12-24 0bf4312
Refactor poetry and add poetry.lock
