+++ .github/workflows/mkdocs.yml
@@ -31,7 +31,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install poetry
-          poetry install 
           poetry install --with dev
       - name: Build docs
         run: make doc
