### 2024-12-24 d3d8612
Refactor pipelines
