+++ .github/workflows/unit-tests.yml
@@ -36,8 +36,7 @@ jobs:
       - name: Install Python dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install poetry
-          poetry install --with dev
+          pip install .[dev]
 
       - name: Download datasets
         run: python -c "from river import datasets; datasets.CreditCard().download(); datasets.Elec2().download(); datasets.Keystroke().download()"
