+++ .github/workflows/mkdocs.yml
@@ -30,8 +30,8 @@ jobs:
       - name: Install Python dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install poetry
-          poetry install --with dev
+          pip install .[dev]
+
       - name: Build docs
         run: make doc
 
