### 2024-12-24 6502ce2
Refactor pipelines
