+++ .github/workflows/unit-tests.yml
@@ -42,6 +42,6 @@ jobs:
         run: python -c "from river import datasets; datasets.CreditCard().download(); datasets.Elec2().download(); datasets.Keystroke().download()"
 
       - name: pytest
-        run: pytest --cov=deep_river -m "not datasets"
+        run: python -m pytest --cov=deep_river -m "not datasets"
       - name: Upload coverage reports to Codecov with GitHub Action
         uses: codecov/codecov-action@v2
