### 2024-12-24 c3ceb9a
Refactor pipelines
