### 2024-12-24 2ed7677
Refactor pipelines
