+++ .github/workflows/unit-tests.yml
@@ -37,7 +37,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install .[dev]
-          pip install pytest, pytest-cov
+          pip install pytest pytest-cov
 
       - name: Download datasets
         run: python -c "from river import datasets; datasets.CreditCard().download(); datasets.Elec2().download(); datasets.Keystroke().download()"
