### 2024-12-24 09ebbf7
Refactor pipelines
