+++ .github/workflows/mkdocs.yml
@@ -30,6 +30,7 @@ jobs:
       - name: Install Python dependencies
         run: |
           python -m pip install --upgrade pip
+          pip install .
           pip install .[dev]
 
       - name: Build docs
