+++ .github/workflows/unit-tests.yml
@@ -36,6 +36,7 @@ jobs:
       - name: Install Python dependencies
         run: |
           python -m pip install --upgrade pip
+          pip install .
           pip install .[dev]
           pip install pytest pytest-cov
 
