### 2024-12-24 99b11bd
Refactor pipelines
