+++ .github/workflows/unit-tests.yml
@@ -38,6 +38,7 @@ jobs:
           python -m pip install --upgrade pip
           pip install .
           pip install .[dev]
+          pip install graphviz
           pip install pytest pytest-cov
 
       - name: Download datasets
