+++ .github/workflows/unit-tests.yml
@@ -25,26 +25,40 @@ jobs:
         with:
           python-version: ${{ matrix.python }}
 
-      - name: Cache Python dependencies
-        uses: actions/cache@v2
+            # install & configure poetry
+      - name: Install Poetry
+        uses: snok/install-poetry@v1
         with:
-          path: ~/.cache/pip
-          key: ${{ runner.os }}-${{ matrix.python }}-pip-${{ hashFiles('**/setup.py') }}
-          restore-keys: |
-            ${{ runner.os }}-${{ matrix.python }}-pip-
+          python-version: ${{ matrix.python }}
+          virtualenvs-create: true
+          virtualenvs-in-project: true
+          installer-parallel: true
 
-      - name: Install Python dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install .
-          pip install .[dev]
-          pip install graphviz
-          pip install pytest pytest-cov
+            # load cached venv if cache exists
+      - name: Load cached venv
+        id: cached-poetry-dependencies
+        uses: actions/cache@v3
+        with:
+          path: .venv
+          key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
+
+      # install dependencies if cache does not exist
+      - name: Install dependencies
+        if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+        run: poetry install --no-interaction --no-root
+
+      # install root project
+      - name: Install project
+        run: poetry install --no-interaction
 
       - name: Download datasets
         run: python -c "from river import datasets; datasets.CreditCard().download(); datasets.Elec2().download(); datasets.Keystroke().download()"
 
-      - name: pytest
-        run: python -m pytest --cov=deep_river -m "not datasets"
+      # run test suite
+      - name: Run tests
+        run: |
+          export PYTHONPATH=$PWD
+          poetry run pytest -v --cov=deep_river -m "not datasets"
+
       - name: Upload coverage reports to Codecov with GitHub Action
         uses: codecov/codecov-action@v2
