+++ .github/workflows/unit-tests.yml
@@ -25,7 +25,6 @@ jobs:
         with:
           python-version: ${{ matrix.python }}
 
-            # install & configure poetry
       - name: Install Poetry
         uses: snok/install-poetry@v1
         with:
@@ -34,34 +33,28 @@ jobs:
           virtualenvs-in-project: true
           installer-parallel: true
 
-            # load cached venv if cache exists
       - name: Load cached venv
         id: cached-poetry-dependencies
         uses: actions/cache@v3
         with:
           path: .venv
-          key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
+          key: venv-${{ runner.os }}-${{ matrix.python }}-${{ hashFiles('**/poetry.lock') }}
 
-      # install dependencies if cache does not exist
       - name: Install dependencies
         if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
         run: poetry install --no-interaction --no-root
 
-      # install root project
       - name: Install project
         run: poetry install --no-interaction
 
       - name: Download datasets
         run: python -c "from river import datasets; datasets.CreditCard().download(); datasets.Elec2().download(); datasets.Keystroke().download()"
-        with:
-          path: .venv
-          key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
 
-      # run test suite
       - name: Run tests
         run: |
           export PYTHONPATH=$PWD
           poetry run pytest -v --cov=deep_river -m "not datasets"
+        shell: bash
 
       - name: Upload coverage reports to Codecov with GitHub Action
         uses: codecov/codecov-action@v2
