+++ .github/workflows/unit-tests.yml
@@ -48,7 +48,7 @@ jobs:
         run: poetry install --no-interaction
 
       - name: Download datasets
-        run: python -c "from river import datasets; datasets.CreditCard().download(); datasets.Elec2().download(); datasets.Keystroke().download()"
+        run: poetry run python -c "from river import datasets; datasets.CreditCard().download(); datasets.Elec2().download(); datasets.Keystroke().download()"
 
       - name: Run tests
         run: |
