+++ .github/workflows/unit-tests.yml
@@ -1,3 +1,13 @@
+name: unit-tests
+
+on:
+  push:
+    branches:
+      - master
+  pull_request:
+    branches:
+      - master
+
 jobs:
   test:
     runs-on: ${{ matrix.os }}
