+++ pyproject.toml
@@ -32,7 +32,7 @@ scikit-learn = "~1.5.0"
 torch = "~2.2.2"
 pandas = "~2.2.2"
 numpy = "~1.26.4"
-river = "~0.21.2"
+river = "*"
 scipy = "~1.13.0"
 tqdm = "~4.66.5"
 ordered-set = "~4.1.0"
