+++ pyproject.toml
@@ -27,7 +27,7 @@ classifiers = [
     "Programming Language :: Python :: Implementation :: PyPy",
 ]
 [tool.poetry.dependencies]
-python = "<4.0,>=3.9"
+python = ">=3.9,<3.13"
 scikit-learn = "~1.5.0"
 torch = "~2.2.2"
 pandas = "~2.2.2"
