+++ .github/workflows/unit-tests.yml
@@ -14,9 +14,9 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python: ['3.9', '3.10', '3.11', '3.12']
+        python: ['3.10', '3.11', '3.12']  # Start testing with Python >=3.10
         os: [ubuntu-latest, macos-latest, windows-latest]
-        river: ['0.21.1', '0.21.2', 'latest']  # Add multiple river versions
+        river: ['0.21.1', '0.22.0', 'latest']
 
     steps:
       - uses: actions/checkout@v4
@@ -46,6 +46,13 @@ jobs:
         run: poetry config virtualenvs.in-project true
         shell: powershell
 
+      - name: Check compatibility
+        run: |
+          if [[ "${{ matrix.python }}" == "3.10" && "${{ matrix.river }}" == "0.21.1" ]]; then
+            echo "Skipping incompatible combination: Python ${{ matrix.python }} and river ${{ matrix.river }}"
+            exit 0
+          fi
+
       - name: Load cached venv
         id: cached-poetry-dependencies
         uses: actions/cache@v3
@@ -65,6 +72,9 @@ jobs:
             poetry add river==${{ matrix.river }}
           fi
         shell: bash
+
+      - name: Debug Poetry Environment
+        run: poetry show --tree
 
       - name: Install project
         run: poetry install --no-interaction
