+++ .github/workflows/unit-tests.yml
@@ -16,7 +16,7 @@ jobs:
       matrix:
         python: ['3.10', '3.11', '3.12']  # Start testing with Python >=3.10
         os: [ubuntu-latest, macos-latest, windows-latest]
-        river: ['0.22.0', 'latest']
+        river: ['0.21.2' , '0.22.0']
 
     steps:
       - uses: actions/checkout@v4
