+++ pyproject.toml
@@ -33,7 +33,6 @@ torch = "~2.2.2"
 pandas = "~2.2.2"
 numpy = "~1.26.4"
 river = "*"
-scipy = "~1.13.0"
 tqdm = "~4.66.5"
 ordered-set = "~4.1.0"
 torchviz = "~0.0.2"
