+++ .github/workflows/mkdocs.yml
@@ -12,29 +12,29 @@ jobs:
       - name: Install Ubuntu dependencies
         run: sudo apt-get install graphviz pandoc
 
-      - name: Set up Python
-        uses: actions/setup-python@v2
+      - name: Install Poetry
+        uses: snok/install-poetry@v1
         with:
-          python-version: 3.12.8
+          python-version: 3.12
+          virtualenvs-create: true
+          virtualenvs-in-project: true
+          installer-parallel: true
 
-      - run: curl https://sh.rustup.rs -sSf | sh -s -- --default-toolchain=nightly --profile=minimal -y && rustup show
-        if: matrix.os == 'ubuntu-latest'
-
-      - name: Cache Python dependencies
-        uses: actions/cache@v2
+      - name: Load cached venv
+        id: cached-poetry-dependencies
+        uses: actions/cache@v3
         with:
-          path: ~/.cache/pip
-          key: ${{ runner.os }}-pip-${{ hashFiles('**/setup.py') }}
-          restore-keys: |
-            ${{ runner.os }}-pip-
-      - name: Install Python dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install .
-          pip install .[dev]
+          path: .venv
+
+      - name: Install dependencies with river
+        run: poetry install --no-interaction
+        shell: bash
 
       - name: Build docs
-        run: make doc
+        run: |
+          cd benchmarks
+          poetry run python render.py
+          mkdocs build
 
       - name: Deploy docs
         run: mkdocs gh-deploy --force
