### 2024-12-28 fd5b052
Add authors' names and emails into pyproject.toml file.
