+++ pyproject.toml
@@ -7,7 +7,10 @@ name = "deep_river"
 version = "2.8.0"  # Match with your versioning system
 description = "Online Deep Learning for river"
 authors = ["Cedric Kulbach <cedric.kulbach@googlemail.com>",
-
+           "Lucas Cazzonelli <Cazzonelli@fzi.de>",
+           "Hoang-Anh Ngo <hoang-anh.ngo@alumni.polytechnique.org>",
+           "Max Halford <maxhalford25@gmail.com>",
+           "Saulo Martiello Mastelini <saulomastelini@gmail.com>",
 ]
 license = "BSD-3-Clause"
 readme = "README.md"
