### 2024-12-28 ba31620
Refactor CONTRIBUTING.md file
