### 2024-12-30 #7226 @hoanganhngo610
Hi @musabgultekin! Based on your concerns and recommendations, we have updated our repository as follows:

- We have added a contributing guideline ([CONTRIBUTING.md](https://github.com/online-ml/deep-river/blob/master/CONTRIBUTING.md)) file.
- We have also incorporated all of the most recent changes into the [newest release (v0.2.8)](https://github.com/online-ml/deep-river/releases/tag/v0.2.8). This version is also now [available on PyPI](https://pypi.org/project/deep-river/) for installation with pip.

Along with the previous addition of the [Related Work](https://github.com/online-ml/deep-river/blob/paper/paper/paper.md#related-work) section to the paper, we hope that our updates have been satisfactory to you. Once again, thank you very much for your valuable reviews towards our paper!

Best regards,
Hoang-Anh
