### 2024-12-30 #7226 @musabgultekin
@hoanganhngo610 , After these changes, It looks great on my end! Thank you very much!
