### 2024-12-30 #7226 @hoanganhngo610
Dear @musabgultekin,

We really appreciate that you find our changes satisfactory. Once again, I would want to send our sincerest thanks to you and @atanikan for your effort towards reviewing the submission with very valuable feedback!

Dear @cheginit, I believe that we have resolved all concerns of both reviewers with the current version. The `paper` branch of our repository is also fullly updated, incorporating all changes from the `master` branch.

Best regards,
Hoang-Anh
