+++ deep_river/preprocessing/__init__.py
@@ -1,4 +1,3 @@
-from deep_river.preprocessing.embedding import (EmbeddingTransformer,
-                                                WordEmbeddingModel)
+from deep_river.preprocessing.embedding import EmbeddingTransformer, WordEmbeddingModel
 
 __all__ = ["EmbeddingTransformer", "WordEmbeddingModel"]
