+++ pyproject.toml
@@ -29,6 +29,7 @@ classifiers = [
 python = ">=3.10,<3.13"
 scikit-learn = "~1.5.0"
 torch = "~2.2.2"
+torchtext = "~0.17.2"
 pandas = "~2.2.2"
 numpy = "~1.26.4"
 river = "*"
