### 2025-01-01 ec2e55e
Update bib entries in JOSS papers
