### 2025-01-01 6fde2cd
Add pair of grave accents for scikit-multiflow, creme and River
 The use of pairs of grave accents for MOA and CapyMOA is not needed, since they are not frequently used in previous literature.
