+++ paper/paper.md
@@ -40,7 +40,7 @@ This adaptation is achieved through online learning or continuous ML technologie
 Although deep learning technologies have demonstrated outstanding performance on predefined datasets, their application to online, streaming, and continuous learning scenarios has been limited.
 
 `DeepRiver`[@githubGitHubOnlinemldeepriver] is a Python package for deep learning on data streams.
-Built on top of River[@montiel2021river] and PyTorch[@paszke2017automatic], it offers a unified API for both supervised and unsupervised learning.
+Built on top of `River`[@montiel2021river] and PyTorch[@paszke2017automatic], it offers a unified API for both supervised and unsupervised learning.
 Additionally, it provides a suite of tools for preprocessing data streams and evaluating deep learning models.
 
 # Statement of need
@@ -51,7 +51,7 @@ To address this imperative, adopting online learning and continuous ML technolog
 While deep learning technologies have demonstrated exceptional performance on static, predefined datasets, their application to dynamic and continuously evolving data streams remains underexplored.
 The absence of widespread integration of deep learning into online, streaming, and continuous learning scenarios hampers the full potential of these advanced algorithms in real-time decision-making [@kulbach2024retrospectivetutorialopportunitieschallenges].
 The emergence of the `DeepRiver`[@githubGitHubOnlinemldeepriver] Python package fills a critical void in the field of deep learning on data streams.
-Leveraging the capabilities of River[@montiel2021river] and PyTorch[@paszke2017automatic], `DeepRiver` offers a unified API for both supervised and unsupervised learning, providing a seamless bridge between cutting-edge deep learning techniques and the challenges posed by dynamic data streams.
+Leveraging the capabilities of `River`[@montiel2021river] and PyTorch[@paszke2017automatic], `DeepRiver` offers a unified API for both supervised and unsupervised learning, providing a seamless bridge between cutting-edge deep learning techniques and the challenges posed by dynamic data streams.
 Moreover, the package equips practitioners with essential tools for data stream preprocessing and the evaluation of deep learning models in dynamic, real-time environments.
 This was already made use of in the context of Streaming Anomaly Detection [@cazzonelli2022detecting].
 As the demand for effective and efficient adaptation of machine learning systems to evolving data structures continues to grow, the integration of `DeepRiver` into the landscape becomes crucial.
@@ -63,19 +63,19 @@ The need for such advancements is evident in the quest to harness the full power
 Online machine learning involves updating models incrementally as new data arrives, rather than retraining models from scratch. 
 Several frameworks and libraries have been developed to support this paradigm:
 
-* scikit-multiflow [@JMLR:v19:18-251]
+* `scikit-multiflow` [@JMLR:v19:18-251]
     * Python-based Library: Inspired by the Java-based MOA framework, designed for streaming data and online learning in Python. 
     * Key Features:
       * Supports algorithms like Hoeffding Trees, online bagging, and boosting. 
       * Includes concept drift detection (e.g., ADWIN, Page-Hinkley) to adapt to changing data distributions. 
       * Stream generators and evaluators for real-time data simulation and model assessment. 
     * Limitations: Focuses mainly on traditional machine learning methods, with limited support for deep learning architectures.
-* creme [@creme]
+* `creme` [@creme]
   * Lightweight Online Learning: Specialized in incremental learning where models are updated per instance, leading to efficient, low-latency model training.
   * Provides a unified API with a broad range of online learning algorithms, making it the go-to library for streaming data analysis in Python.
   Limitations: Primarily supports feature-based models with limited capabilities for deep neural networks.
 
[…]
