### 2025-01-01 a5c1101
Minor refactor for the paper
