### 2025-01-01 00f3e58
Minor refactor for the paper
