### 2025-01-01 #7226 @cheginit
@hoanganhngo610 and @kulbachcedric, please address the following in the paper:

- [ ] Since the submission is named "Deep-River" please use the same in the paper too, i.e., replace all instances of "DeepRiver" with "Deep-River" to match the submission title.
- [ ] **Lines 8-10**: The phrase "Machine learning algorithms enhance and expedite decision-making processes based on available data." could be simplified for conciseness.

- [ ] **Lines 20-21**: The sentence "The acceleration facilitated by these algorithms necessitates constant adaptation to evolving data structures, as what is valid at one moment may no longer hold in the future." is verbose and could be rewritten for clarity.

- [ ] **Lines 35-36**: "This was already made use of in the context of Streaming Anomaly Detection" has an awkward structure and could be rephrased for better readability.

- [ ] **Lines 98-99**: The abbreviation "i.ex." is incorrect; it should be "e.g.".

- [ ] **Lines 106-107**: The phrase "Further, it integrates the huge functionality of PyTorch" could be rephrased to "Additionally, it incorporates PyTorch's extensive functionality" for improved style and readability.

- [ ] **Lines 114-115**: The sentence "It provides comprehensive documentation to guide users through the installation, implementation, and customization processes." is disconnected from the preceding text. Consider integrating it more smoothly.

- [ ] **Lines 121-122**: The phrase "Further, it enables the model to adapt to changes in the number of classes over time for classification tasks." is repetitive in its use of "Further." Consider varying transitions.

- [ ] **Lines 129-130**: The sentence "Each category contains several specialized classes that inherit from more general base classes, forming a hierarchical structure." is redundant given the detailed examples that follow.

- [ ] **Lines 145-146**: The **Acknowledgements** section is too brief. Specify the exact nature of support from the Research Center for Information Technology for completeness, i.e., if it's an internal grant, say so, or if it's from a funding agency you need to provide a grant number or something similar.
- [ ] **Lines 14 and 92**: There's no need to reference Deep-River's Github repo in the text, please remove all its instances. Also, remove it from References (Line 159)
- [ ] **Line 170**: Capitalize Python.
- [ ] **Line 175**: Use the correct citation from [here](https://www.semanticscholar.org/paper/Automatic-differentiation-in-PyTorch-Paszke-Gross/b36a5bb1707bb9c70025294b3a310138aae8327a).
