+++ deep_river/anomaly/ae.py
@@ -13,8 +13,9 @@ from deep_river.utils.tensor_conversion import df2tensor
 
 
 class _TestAutoencoder(torch.nn.Module):
-    def __init__(self, n_features, latent_dim=3):
+    def __init__(self, latent_dim=3):
         super().__init__()
+        n_features = 2
         self.linear1 = nn.Linear(n_features, latent_dim)
         self.nonlin = torch.nn.LeakyReLU()
         self.linear2 = nn.Linear(latent_dim, n_features)
@@ -28,31 +29,26 @@ class _TestAutoencoder(torch.nn.Module):
 
 class Autoencoder(DeepEstimator, AnomalyDetector):
     """
-    Wrapper for PyTorch autoencoder models that uses the networks
+    Wrapper for PyTorch autoencoder models that uses the network's
     reconstruction error for scoring the anomalousness of a given example.
 
     Parameters
     ----------
-    module
-        Torch Module that builds the autoencoder to be wrapped.
-        The Module should accept parameter `n_features` so that the returned
-        model's input shape can be determined based on the number of features
-        in the initial training example.
-    loss_fn
-        Loss function to be used for training the wrapped model. Can be a
-        loss function provided by `torch.nn.functional` or one of the
-        following: 'mse', 'l1', 'cross_entropy', 'binary_crossentropy',
-        'smooth_l1', 'kl_div'.
-    optimizer_fn
-        Optimizer to be used for training the wrapped model. Can be an
-        optimizer class provided by `torch.optim` or one of the following:
-        "adam", "adam_w", "sgd", "rmsprop", "lbfgs".
-    lr
-        Learning rate of the optimizer.
-    device
-        Device to run the wrapped model on. Can be "cpu" or "cuda".
-    seed
-        Random seed to be used for training the wrapped model.
+    module : torch.nn.Module
+        Fully initialized PyTorch autoencoder model.
+    loss_fn : Union[str, Callable]
+        Loss function for training. Can be one of the predefined loss functions
+        ('mse', 'l1', etc.) or a callable.
+    optimizer : Union[str, Callable]
+        Optimizer for training. Can be a predefined optimizer ('adam', 'sgd', etc.) or a callable.
+    lr : float
+        Learning rate for the optimizer.
+    is_feature_incremental : bool
+        Whether to adapt to new input features dynamically.
+    device : str
+        Device to run the model on ('cpu' or 'cuda').
+    seed : int
+        Random seed for reproducibility.
     **kwargs
         Parameters to be passed to the `torch.Module` class
         aside from `n_features`.
@@ -71,8 +67,9 @@ class Autoencoder(DeepEstimator, AnomalyDetector):
     >>> metric = metrics.RollingROCAUC(window_size=5000)
 
     >>> class MyAutoEncoder(torch.nn.Module):
-    ...     def __init__(self, n_features, latent_dim=3):
+    ...     def __init__(self, latent_dim=3):
     ...         super(MyAutoEncoder, self).__init__()
+    ...         n_features = 2
     ...         self.linear1 = nn.Linear(n_features, latent_dim)
     ...         self.nonlin = torch.nn.LeakyReLU()
     ...         self.linear2 = nn.Linear(latent_dim, n_features)
@@ -84,7 +81,7 @@ class Autoencoder(DeepEstimator, AnomalyDetector):
     ...         X = self.linear2(X)
     ...         return self.sigmoid(X)
 
-    >>> ae = Autoencoder(module=MyAutoEncoder, lr=0.005)
+    >>> ae = Autoencoder(module=MyAutoEncoder(), lr=0.005)
     >>> scaler = MinMaxScaler()
[…]
