+++ benchmarks/config.py
@@ -35,7 +35,7 @@ MODELS = {
             | TorchClassifier(
                 module=TorchLogisticRegression,
                 loss_fn="binary_cross_entropy",
-                optimizer_fn="sgd",
+                optimizer="sgd",
                 is_class_incremental=True,
                 lr=LEARNING_RATE,
             )
@@ -45,7 +45,7 @@ MODELS = {
             | TorchClassifier(
                 module=TorchMLPClassifier,
                 loss_fn="binary_cross_entropy",
-                optimizer_fn="sgd",
+                optimizer="sgd",
                 is_class_incremental=True,
                 lr=LEARNING_RATE,
             )
@@ -55,7 +55,7 @@ MODELS = {
             | TorchRollingClassifier(
                 module=TorchLSTMClassifier,
                 loss_fn="binary_cross_entropy",
-                optimizer_fn="sgd",
+                optimizer="sgd",
                 is_class_incremental=True,
                 lr=LEARNING_RATE,
                 window_size=20,
@@ -71,7 +71,7 @@ MODELS = {
             | TorchRegressor(
                 module=TorchLinearRegression,
                 loss_fn="mse",
-                optimizer_fn="sgd",
+                optimizer="sgd",
                 lr=LEARNING_RATE,
             )
         ),
@@ -80,7 +80,7 @@ MODELS = {
             | TorchRegressor(
                 module=TorchMLPRegressor,
                 loss_fn="mse",
-                optimizer_fn="sgd",
+                optimizer="sgd",
                 lr=LEARNING_RATE,
             )
         ),
@@ -100,7 +100,7 @@ MODELS = {
             | TorchRollingRegressor(
                 module=TorchLSTMRegressor,
                 loss_fn="mse",
-                optimizer_fn="sgd",
+                optimizer="sgd",
                 lr=LEARNING_RATE,
                 window_size=20,
                 append_predict=False,

[…]
