### 2025-01-02 f101c42
add forcasting method
