### 2025-01-04 05f1075
Remove DeepRiver's GitHub reference and its appearances at lines 14, 28, 92, 97, 104, 119 within the paper
