+++ paper/paper.bib
@@ -91,10 +91,3 @@
 	year = {2024},
 	note = {[Accessed 18-11-2024]},
 }
-@misc{githubGitHubOnlinemldeepriver,
-	author = {},
-	title = {{G}it{H}ub --- online-ml/deep-river --- github.com},
-	howpublished = {\url{https://github.com/online-ml/deep-river}},
-	year = {},
-	note = {[Accessed 18-11-2024]},
-}
