+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{montiel2021river,
-  title={River: machine learning for streaming data in Python},
+  title={River: machine learning for streaming data in {P}ython},
   author={Montiel, Jacob and Halford, Max and Mastelini, Saulo Martiello
           and Bolmier, Geoffrey and Sourty, Raphael and Vaysse, Robin and Zouitine, Adil
           and Gomes, Heitor Murilo and Read, Jesse and Abdessalem, Talel and Bifet, Albert},
