### 2025-01-04 e05c4b0
Update Paszke's Auutomatic differentiation in Pytorch paper with correct bib entry from editor's suggestion
