+++ paper/paper.bib
@@ -31,9 +31,9 @@
 }
 @inproceedings{paszke2017automatic,
   title={Automatic differentiation in PyTorch},
-  author={Paszke, Adam and Gross, Sam and Chintala, Soumith and Chanan, Gregory and Yang, Edward and DeVito, Zachary and Lin, Zeming and Desmaison, Alban and Antiga, Luca and Lerer, Adam},
-  booktitle={NIPS-W},
-  year={2017}
+  author={Paszke, Adam and Gross, Sam and Chintala, Soumith and Chanan, Gregory and Yang, Edward and DeVito, Zach and Lin, Zeming and Desmaison, Alban and Antiga, Luca and Lerer, Adam},
+  year={2017},
+  url={https://api.semanticscholar.org/CorpusID:40027675}
 }
 @article{bayram2022concept,
   title={From concept drift to model degradation: An overview on performance-aware drift detectors},
