### 2025-01-04 81e8eec
Remove redundant sentence from paper
Remove the sentence "Each category contains several specialized classes that inherit from more general base classes, forming a hierarchical structure.", as itis redundant given the detailed examples below.
