### 2025-01-04 a78b2b6
Change the sentence to avoid repetition in the use of the word Further
Change the sentence "Further, it enables the model to adapt ..." to "Moreover, it allows the model to dynamically adjust ..."
