### 2025-01-04 405dce5
Simplify the first sentence of Summary section for conciseness.
Simplify the sentence "Machine learning algorithms enhance and expedite decision-making processes based on available data." to "Machine learning algorithms enhance decision-making deficiency by leveraging available data.
