+++ paper/paper.md
@@ -104,7 +104,7 @@ This addresses a critical gap left by existing frameworks, which are predominant
 # Features
 
 `DeepRiver` enables the usage of deep learning models for data streams.
-This means that deep learning models need to adapt to changes within the evolving data stream [@bayram2022concept;@lu2018learning] i.ex. the number of classes might change over time.
+This means that deep learning models need to adapt to changes within the evolving data stream [@bayram2022concept;@lu2018learning] e.g. the number of classes might change over time.
 In addition to the integration of PyTorch[@paszke2017automatic] into `River`[@montiel2021river], this package offers additional data stream specific functionalities such as class incremental learning or specific optimizers for data streams.
 
 ## Compatibility
