### 2025-01-05 dbfdfe6
Modify sentence in line 35-36 for better readability
Modify the sentence "This was already made use of in the context of Streaming Anomaly Detection." into "This functionality has been applied to Streaming Anomaly Detection."
