### 2025-01-05 7110359
Modify sentence in line 20-21 for clarity
Change the sentence "The acceleration facilitated by these algorithms necessitates adaptation to evolving data structures, as what is valid at one moment may no longer hold in the future." to "These algorithms, while accelerating analysis, require continuous adaptation to dynamic data structures, as patters may evolve rapidly."
