### 2025-01-05 810715b
Modify sentence line 114-115 to connect with the preceding text more smoothly
