+++ paper/paper.md
@@ -113,8 +113,8 @@ In addition to the integration of PyTorch[@paszke2017automatic] into `River`[@mo
 Additionally, it incorporates PyTorch's [@paszke2017automatic] extensive functionality for deep learning such as using GPU acceleration and a broad range of architectures.
 This unified approach simplifies the development process and facilitates a cohesive workflow for practitioners working with dynamic data streams.
 Leveraging the capabilities of the well-established `River`[@montiel2021river] library and the powerful PyTorch[@paszke2017automatic] framework, `DeepRiver` combines the strengths of these technologies to deliver a robust and flexible platform for deep learning on data streams.
-This foundation ensures reliability, scalability, and compatibility with state-of-the-art machine learning methodologies.
-It provides comprehensive [documentation](https://online-ml.github.io/deep-river/) to guide users through the installation, implementation, and customization processes. Additionally, a supportive community ensures that users have access to resources, discussions, and assistance, fostering a collaborative environment for continuous improvement and knowledge sharing.
+This foundation ensures reliability, scalability, and compatibility with state-of-the-art machine learning methodologies, with comprehensive [documentation](https://online-ml.github.io/deep-river/) guiding users through the installation, implementation, and customization processes.
+Additionally, a supportive community ensures that all `DeepRiver`'s users have access to resources, discussions, and assistance, fostering a collaborative environment for continuous improvement and knowledge sharing.
 
 ## Adaptivity
 
