### 2025-01-05 #7226 @hoanganhngo610
Hi @cheginit (pinging @kulbachcedric for future follow-up),

First of all, thank you very much for a very detailed final checks for out submission. We really appreciate it! We have tried to address all of your suggestions within the PR [#107](https://github.com/online-ml/deep-river/pull/107/files), and we hope that the changes made are satisfactory. Please find below a point-by-point response below:

>  Since the submission is named "Deep-River" please use the same in the paper too, i.e., replace all instances of "DeepRiver" with "Deep-River" to match the submission title.

Originally, at the time of submission, and on previous publications/tutorials related to `DeepRiver`, we used the name `Deep-River`. However, after several discussion and by consensus, we decided that it would be best to go for the use of the name `DeepRiver` for consistency with other packages within the `River` ecosystem, for example `LightRiver`. As such, if possible, we would really appreciate if we can retain the use of the name `DeepRiver` within our paper.

> **Lines 8-10**: The phrase "Machine learning algorithms enhance and expedite decision-making processes based on available data." could be simplified for conciseness.

We have modified this sentence into "Machine learning algorithms enhance decision-making efficiency by leveraging available data."

> **Lines 20-21**: The sentence "The acceleration facilitated by these algorithms necessitates constant adaptation to evolving data structures, as what is valid at one moment may no longer hold in the future." is verbose and could be rewritten for clarity.

We have rewritten this sentence into "These algorithms, while accelerating analysis, require continuous adaptation to dynamic data structures, as patterns may evolve rapidly."

> **Lines 35-36**: "This was already made use of in the context of Streaming Anomaly Detection" has an awkward structure and could be rephrased for better readability.

We have changed the sentence into "Such functionality has been applied to Streaming Anomaly Detection."

> **Lines 98-99**: The abbreviation "i.ex." is incorrect; it should be "e.g.".

We have made the change accordingly.

> **Lines 106-107**: The phrase "Further, it integrates the huge functionality of PyTorch" could be rephrased to "Additionally, it incorporates PyTorch's extensive functionality" for improved style and readability.

We have made the change accordingly.

>  **Lines 114-115**: The sentence "It provides comprehensive documentation to guide users through the installation, implementation, and customization processes." is disconnected from the preceding text. Consider integrating it more smoothly.

We have modified the last sentences of the paragraph to make this sentence more connected to the preceding texts, as follows:

"This foundation ensures reliability, scalability, and compatibility with state-of-the-art machine learning methodologies, with comprehensive [documentation](https://online-ml.github.io/deep-river/) guiding users through the installation, implementation, and customization processes. Additionally, a supportive community ensures that all `DeepRiver`'s users have access to resources, discussions, and assistance, fostering a collaborative environment for continuous improvement and knowledge sharing."

> **Lines 121-122**: The phrase "Further, it enables the model to adapt to changes in the number of classes over time for classification tasks." is repetitive in its use of "Further." Consider varying transitions.

We have adjusted the sentence into "Furthermore, it allows the model to dynamically adjust to changes in the number of classes over time for classification tasks."

> **Lines 129-130**: The sentence "Each category contains several specialized classes that inherit from more general base classes, forming a hierarchical structure." is redundant given the detailed examples that follow.

We have removed the sentence accordingly.

>  **Lines 145-146**: The **Acknowledgements** section is too brief. Specify the exact nature of support from the Research Center for Information Technology for completeness, i.e., if it's an internal grant, say so, or if it's from a funding agency you need to provide a grant number or something similar.

The original **Acknowledgements** section is intended to only indicate that Cedric Kulbach and Lucas Cazzonelli participated in the development of `DeepRiver` when both were working at FZI Research Center for Information Technology. However, there is no direct financial support from the Research Center for this work.

This part has been modified to showcase that I (Hoang-Anh Ngo) received an External Study Awards from the AI Institute, University of Waikato to conduct research on onlineML, including participating in the development of `DeepRiver`.

> **Lines 14 and 92**: There's no need to reference Deep-River's Github repo in the text, please remove all its instances. Also, remove it from References (Line 159)

We have deleted the bib entry for DeepRiver's GitHub repo accordingly, and remove all instances appearing in lines 14, 28, 92, 97, 104, 119.

> **Line 170**: Capitalize Python.

We have made the change accordingly.

> **Line 175**: Use the correct citation from [here](https://www.semanticscholar.org/paper/Automatic-differentiation-in-PyTorch-Paszke-Gross/b36a5bb1707bb9c70025294b3a310138aae8327a).

We have updated the citation in our bib file with the citation of this paper from Semantic Scholar accordingly.
