### 2025-01-05 12df82e
Fix typo within JOSS paper
