### 2025-01-05 #7226 @cheginit
@hoanganhngo610 Thanks for working on addressing the comments:

> As such, if possible, we would really appreciate if we can retain the use of the name DeepRiver within our paper.

The main point is that you need to be consistent with using DeepRiver vs Deep-River. Since the submission title and the archive that you will need to create later should match. So please make sure to use DeepRiver when creating the archive. I changed the submission title. So use the same title for the archive.
