### 2025-01-05 #7226 @cheginit
@kulbachcedric is v0.2.6 the final version that you want to be associated with your JOSS submission? If not, please release a new version, then give me the new version. Once you took care of the version, create an archive, e.g., using Zenodo, making sure to use the same exact title as the paper and your submission and provide me with the DOI of the archive.
