### 2025-01-05 #7226 @hoanganhngo610
Hi @cheginit,

> The main point is that you need to be consistent with using DeepRiver vs Deep-River. Since the submission title and the archive that you will need to create later should match. So please make sure to use DeepRiver when creating the archive. I changed the submission title. So use the same title for the archive.

Thank you very much for your reminder. We will be consistent with the use of `DeepRiver` from now on, especially with this publication and when creating the archive on Zenodo.

> @kulbachcedric is v0.2.6 the final version that you want to be associated with your JOSS submission? If not, please release a new version, then give me the new version. Once you took care of the version, create an archive, e.g., using Zenodo, making sure to use the same exact title as the paper and your submission and provide me with the DOI of the archive.

After incorporating all changes suggested by the reviewers, we have released [version v0.2.8](https://github.com/online-ml/deep-river/releases/tag/v0.2.8) recently. We would want this version to be associated with this JOSS submission. Once we have finished creating the archive on Zenodo, we will let you know ASAP!
