### 2025-01-05 #7226 @hoanganhngo610
Hi @cheginit, we have uploaded the latest version (v0.2.8) onto [Zenodo](https://doi.org/10.5281/zenodo.14601979).

If there is any other tasks that need to be done, please do not hesitate to let us know!
