### 2025-01-05 #7226 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
