### 2025-01-06 #7226 @crvernon
@kulbachcedric and @hoanganhngo610 - I just need you to address the following before I move to accept this for publication:

- [ ] LINE 15:  There should be a space in between "River" and "(Montiel et al., 2021)"
- [ ] LINE 16  There should be a space in between "PyTorch" and "(Paszke et al., 2017)"
- [ ] LINE 30:  There should be a space in between "River" and "(Montiel et al., 2021)"
- [ ] LINE 30  There should be a space in between "PyTorch" and "(Paszke et al., 2017)"
- [ ] LINE 100:  There should be a space in between "River" and "(Montiel et al., 2021)"
- [ ] LINE 100  There should be a space in between "PyTorch" and "(Paszke et al., 2017)"

This has been very well done and the minimal changes mentioned above will suffice.  Thanks!
