### 2025-01-06 bd5620a
Adding spaces between package names and references in JOSS paper
Adding spaces between package names and references at lines 15, 16, 30, 100 as suggested by the EiC of JOSS.
