### 2025-01-06 #7226 @hoanganhngo610
Hi @crvernon! Thank you so much for your suggestions, and also for your compliment towards our submission! We have addressed all of your suggestions in [our latest commit](https://github.com/online-ml/deep-river/commit/bd5620a4340263000d3aad23c97af95d713099c7) to both the `master` and `paper` branch.

We hope that the changes made are satisfactory!

Best regards,
Hoang-Anh
