### 2025-01-06 #7226 @crvernon
🥳 Congratulations on your new publication @kulbachcedric and @hoanganhngo610! Many thanks to @cheginit for editing and @musabgultekin and @atanikan for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
