### 2025-01-06 #7226 @hoanganhngo610
Thank you so much for your suggestion @crvernon! We agree that we may have been a bit too "concise" within our documentation and the examples within our notebook, and that may also be needed refactoring ASAP in the future.

Moreover, besides, we are also trying our best to host tutorials at different ML conferences/venues with more thorough explanation/examples with the use cases. A quite successful one would the the one we presented at [ECML-PKDD 2023](https://arxiv.org/abs/2405.17222).

Once again, thank you very much for your support and constructive suggestions for our repository/paper! We really appreciate it!
